+++ scripts/build_demand_profiles_from_eia.py
@@ -1,24 +1,34 @@
+# -*- coding: utf-8 -*-
 # SPDX-FileCopyrightText:  Open Energy Transition gGmbH
 #
 # SPDX-License-Identifier: AGPL-3.0-or-later
 
 import os
 import sys
-sys.path.append(os.path.abspath(os.path.join(__file__ ,"../../")))
+
+sys.path.append(os.path.abspath(os.path.join(__file__, "../../")))
 import datetime as dt
-import pandas as pd
+
 import geopandas as gpd
 import numpy as np
+import pandas as pd
 import pypsa
-from scripts._helper import mock_snakemake, update_config_from_wildcards, create_logger, \
-                            configure_logging, get_colors, BASE_PATH
+
+from scripts._helper import (
+    BASE_PATH,
+    configure_logging,
+    create_logger,
+    get_colors,
+    mock_snakemake,
+    update_config_from_wildcards,
+)
 
 logger = create_logger(__name__)
 
 
 def parse_inputs():
     """
-    Load all input data  
+    Load all input data
     Parameters
     ----------
     Returns
@@ -54,7 +64,7 @@ def parse_inputs():
 
 def build_demand_profiles(df_utility_demand, df_ba_demand, gdf_ba_shape, pypsa_network):
     """
-    Build spatiotemporal demand profiles 
+    Build spatiotemporal demand profiles
     Parameters
     ----------
     df_utility_demand: geopandas dataframe
@@ -105,7 +115,9 @@ def build_demand_profiles(df_utility_demand, df_ba_demand, gdf_ba_shape, pypsa_n
     pypsa_gpd["color"] = get_colors(len(pypsa_gpd))
 
     df_utility_centroid = gpd.sjoin_nearest(df_utility_centroid, pypsa_gpd, how="left")
-    df_utility_centroid.rename(columns={"index_right": "PyPSA_bus", "Bus": "PyPSA_bus"}, inplace=True)
+    df_utility_centroid.rename(
+        columns={"index_right": "PyPSA_bus", "Bus": "PyPSA_bus"}, inplace=True
+    )
 
     df_demand_bus = pd.DataFrame(
         index=pd.to_datetime(df_ba_demand.index),
@@ -128,7 +140,9 @@ def build_demand_profiles(df_utility_demand, df_ba_demand, gdf_ba_shape, pypsa_n
     df_demand_bus_timeshifted = df_demand_bus_timeshifted[:8760]
     df_demand_bus_timeshifted.index = pypsa_network.snapshots
     df_demand_bus_timeshifted.index.name = "time"
-    logger.info("Built demand_profiles.csv based on demand distribution using utility level and balancing authority demand data.")
+    logger.info(
+        "Built demand_profiles.csv based on demand distribution using utility level and balancing authority demand data."
+    )
     return df_demand_bus_timeshifted
 
 
@@ -146,7 +160,9 @@ def read_scaling_factor(demand_scenario, horizon):
     scaling_factor: pandas dataframe
         Scaling factor of demand projection scenario
     """
-    horizon = 2024 if horizon == 2025 else horizon # select 2024 demand projection for 2025 horizon
+    horizon = (
+        2024 if horizon == 2025 else horizon
+    )  # select 2024 demand projection for 2025 horizon
     foldername = os.path.join(BASE_PATH, snakemake.params.demand_projections)
     filename = f"Scaling_Factor_{demand_scenario}_Moderate_{horizon}_by_state.csv"
     scaling_factor = pd.read_csv(os.path.join(foldername, filename), sep=";")
@@ -184,7 +200,8 @@ def get_spatial_mapping(pypsa_network, gadm_shape_path, geo_crs):
     # map gadm shapes to each bus
     spatial_gadm_bus_mapping = (
         buses_gdf.sjoin(gadm_shape, how="left", predicate="within")
-        .set_index("Bus")["ISO_1"].str.replace("US-", "")
+        .set_index("Bus")["ISO_1"]
+        .str.replace("US-", "")
     )
     return spatial_gadm_bus_mapping
 
@@ -206,15 +223,21 @@ def scale_demand_profiles(df_demand_profiles, spatial_gadm_bus_mapping, scaling_
          Scaled demand profiles based on the demand projections
     """
     # convert demand_profiles from wide to long format
-    df_demand_long = df_demand_profiles.melt(ignore_index=False, var_name="Bus", value_name="demand")
+    df_demand_long = df_demand_profiles.melt(
+        ignore_index=False, var_name="Bus", value_name="demand"
+    )
 
     # map Bus IDs to State Codes
     df_demand_long["region_code"] = df_demand_long["Bus"].map(spatial_gadm_bus_mapping)
 
     # merge with scaling_factor DataFrame based on region_code and time
     df_demand_long.reset_index(inplace=True)
-    scaling_factor["time"] = scaling_factor["time"].apply(lambda t: t.replace(year=df_demand_long["time"][0].year))
-    df_scaled = df_demand_long.merge(scaling_factor, on=["region_code", "time"], how="left")
+    scaling_factor["time"] = scaling_factor["time"].apply(
+        lambda t: t.replace(year=df_demand_long["time"][0].year)
+    )
+    df_scaled = df_demand_long.merge(
+        scaling_factor, on=["region_code", "time"], how="left"
+    )
     del scaling_factor, df_demand_long
 
     # multiply demand by scaling factor
@@ -222,8 +245,12 @@ def scale_demand_profiles(df_demand_profiles, spatial_gadm_bus_mapping, scaling_
     df_scaled["scaled_demand"] = df_scaled["demand"] * df_scaled["scaling_factor"]
 
     # pivot back to original wide format
-    scaled_demand_profiles = df_scaled.pivot(index="time", columns="Bus", values="scaled_demand")
-    scaled_demand_profiles = scaled_demand_profiles[sorted(scaled_demand_profiles.columns)]
+    scaled_demand_profiles = df_scaled.pivot(
+        index="time", columns="Bus", values="scaled_demand"
+    )
+    scaled_demand_profiles = scaled_demand_profiles[
+        sorted(scaled_demand_profiles.columns)
+    ]
     logger.info(f"Scaled demand based on scaling factor for each state.")
 
     return scaled_demand_profiles
@@ -246,13 +273,17 @@ def read_data_center_profiles(horizon, data_center_profiles_path):
     foldername = os.path.join(BASE_PATH, data_center_profiles_path)
     filename = f"data_center_profile_{horizon}_by_state.csv"
     data_center_profile = pd.read_csv(os.path.join(foldername, filename))
-    statewise_data_center_load = data_center_profile.groupby("region_code")["load_GW"].mean()
+    statewise_data_center_load = data_center_profile.groupby("region_code")[
+        "load_GW"
+    ].mean()
     logger.info(f"Read {filename} for setting data center demands for {horizon}.")
 
     return statewise_data_center_load
 
 
-def add_data_center_demand(df_demand_profiles, spatial_gadm_bus_mapping, data_center_demand):
+def add_data_center_demand(
+    df_demand_profiles, spatial_gadm_bus_mapping, data_center_demand
+):
     """
     Adds data center demand to demand profile
     Parameters
@@ -279,7 +310,9 @@ def add_data_center_demand(df_demand_profiles, spatial_gadm_bus_mapping, data_ce
             # data center demand for the gadm region in GW
             region_data_center_demand = data_center_demand.loc[bus_region]
             # take fraction of regional data center demand
-            bus_data_center_demand = (region_data_center_demand * 1e3) / region_counts.loc[bus_region]
+            bus_data_center_demand = (
+                region_data_center_demand * 1e3
+            ) / region_counts.loc[bus_region]
             # add bus-level data center demand
             df_demand_profiles.loc[:, bus] += bus_data_center_demand
 
@@ -300,7 +333,6 @@ if __name__ == "__main__":
     plot_path = os.path.join(BASE_PATH, "plots/demand_modelling")
     os.makedirs(plot_path, exist_ok=True)
 
-
     df_ba_demand, gdf_ba_shape, df_utility_demand, pypsa_network = parse_inputs()
 
     df_demand_profiles = build_demand_profiles(
@@ -312,13 +344,19 @@ if __name__ == "__main__":
 
     # scale demand for future scenarios
     if snakemake.params.demand_horizon >= 2025:
-        scaling_factor = read_scaling_factor(snakemake.params.demand_scenario, snakemake.params.demand_horizon)
-        df_demand_profiles = scale_demand_profiles(df_demand_profiles, spatial_gadm_bus_mapping, scaling_factor)
+        scaling_factor = read_scaling_factor(
+            snakemake.params.demand_scenario, snakemake.params.demand_horizon
+        )
+        df_demand_profiles = scale_demand_profiles(
+            df_demand_profiles, spatial_gadm_bus_mapping, scaling_factor
+        )
 
     # set data center demand profiles
     if snakemake.config["demand_projection"]["data_centers_load"]:
         data_center_demand = read_data_center_profiles(snakemake.params.demand_horizon, snakemake.params.data_center_profiles)
-        df_demand_profiles = add_data_center_demand(df_demand_profiles, spatial_gadm_bus_mapping, data_center_demand)
+        df_demand_profiles = add_data_center_demand(
+            df_demand_profiles, spatial_gadm_bus_mapping, data_center_demand
+        )
 
     # save demand_profiles.csv
     df_demand_profiles.to_csv(snakemake.output.demand_profile_path)