+++ scripts/retrieve_base_network.py
@@ -1,15 +1,22 @@
+# -*- coding: utf-8 -*-
 # SPDX-FileCopyrightText:  Open Energy Transition gGmbH
 #
 # SPDX-License-Identifier: AGPL-3.0-or-later
 
 import os
 import sys
-sys.path.append(os.path.abspath(os.path.join(__file__ ,"../../")))
-import warnings
-warnings.filterwarnings("ignore")
-from scripts._helper import mock_snakemake, update_config_from_wildcards, create_logger, \
-                            download_and_unzip_gdrive, configure_logging
 
+sys.path.append(os.path.abspath(os.path.join(__file__, "../../")))
+import warnings
+
+warnings.filterwarnings("ignore")
+from scripts._helper import (
+    configure_logging,
+    create_logger,
+    download_and_unzip_gdrive,
+    mock_snakemake,
+    update_config_from_wildcards,
+)
 
 logger = create_logger(__name__)
 
@@ -30,9 +37,9 @@ if __name__ == "__main__":
     config_base_network = config["custom_databundles"]["bundle_base_network_USA"]
 
     # destination for base.nc
-    destination=os.path.dirname(snakemake.output[0])
+    destination = os.path.dirname(snakemake.output[0])
 
     # download base.nc
-    downloaded = download_and_unzip_gdrive(config_base_network,
-                                           destination=destination,
-                                           logger=logger)
+    downloaded = download_and_unzip_gdrive(
+        config_base_network, destination=destination, logger=logger
+    )